Python Code
Some parts I'll need to fix up a bit. I'll work some more on it
on
Saturday.
I'm using the turtle module to draw the grid. I've made the
'Enter
moves' part, but I'll need to add the AI in on Sat as well. I've
done
a basic understanding of move tracking to check if the user goes
back on themselves. If they do, the program ends the game immediately.
I've also added a move numbering system which could be used as a score
if a single
player mode is made for the apps, (so the users could
compare their
score (maybe on a leader-board? :+1: ), the higher the
amount of moves
before dying the better). I'll probably get the AI
sorted out on Sat as I'll have longer after school.
I'll update with
more info tomorrow. :panda_face: :smile: